refactor(wizard): extract height calculation to utility function
