+++ kit/dapp/src/components/multistep-form/multistep-wizard.tsx
@@ -3,6 +3,7 @@ import { cn } from "@/lib/utils";
 import { createLogger } from "@settlemint/sdk-utils/logging";
 import { useForm } from "@tanstack/react-form";
 import { useCallback, useMemo } from "react";
+import { calculateWizardHeight } from "./calculate-wizard-height";
 import type { MultiStepWizardProps, WizardContextValue } from "./types";
 import { useMultiStepWizardState } from "./use-multistep-wizard-state";
 import { WizardProvider } from "./wizard-context";
@@ -223,62 +224,10 @@ export function MultiStepWizard<TFormData = Record<string, unknown>>({
   }, []);
 
   // Calculate dynamic height based on content
-  const dynamicHeight = useMemo(() => {
-    const baseHeight = 300; // Base height for title, progress, etc. (increased)
-    const stepHeight = 100; // Approximate height per step (increased)
-    const groupHeaderHeight = 80; // Height for group headers (increased)
-    const spacingPadding = 100; // Additional padding for margins, spacing, etc.
-    const minHeight = 600; // Minimum height
-    const maxHeight = 1000; // Maximum height to prevent excessive size (increased)
-
-    if (!groups || groups.length === 0) {
-      // No groups, calculate based on total steps
-      return Math.min(
-        Math.max(
-          baseHeight + steps.length * stepHeight + spacingPadding,
-          minHeight
-        ),
-        maxHeight
-      );
-    }
-
-    // Calculate height needed for the largest group when expanded
-    const maxGroupSize =
-      groups.length > 0
-        ? Math.max(
-            ...groups.map((group) => {
-              const groupSteps = steps.filter(
-                (step) => step.groupId === group.id
-              );
-              return groupSteps.length;
-            })
-          )
-        : 0;
-
-    // Calculate total height needed
-    const totalGroupHeaders = groups.length * groupHeaderHeight;
-    const maxGroupContent = maxGroupSize * stepHeight;
-    const calculatedHeight =
-      baseHeight + totalGroupHeaders + maxGroupContent + spacingPadding;
-    const finalHeight = Math.min(
-      Math.max(calculatedHeight, minHeight),
-      maxHeight
-    );
-
-    logger.debug("Dynamic height calculation", {
-      baseHeight,
-      totalGroupHeaders,
-      maxGroupContent,
-      spacingPadding,
-      calculatedHeight,
-      finalHeight,
-      groupsCount: groups.length,
-      maxGroupSize,
-      totalSteps: steps.length,
-    });
-
-    return finalHeight;
-  }, [steps, groups]);
+  const dynamicHeight = useMemo(
+    () => calculateWizardHeight(steps, groups),
+    [steps, groups]
+  );
 
   // NOW handle conditional rendering after all hooks have been called
   if (steps.length === 0) {
